Add files via upload
This repository contains an in-depth analysis for the Kaggle competition "Spaceship Titanic" (https://www.kaggle.com/competitions/spaceship-titanic), which centers on forecasting which passengers were transported to an alternate dimension. Using Python, machine learning techniques, and regression models, the project thoroughly examines the dataset to formulate predictions. Key tasks encompass data preprocessing, feature engineering, and deploying advanced predictive algorithms. The primary goal is to generate insights and predictions that contribute to the global ranking leaderboard, based on the accuracy of test data validation. Beyond achieving a high leaderboard placement, the project aims to offer a comprehensive understanding of the factors influencing the observed transportation phenomenon aboard the Spaceship Titanic.